+++ babel.config.js
@@ -10,9 +10,6 @@ module.exports = {
     '@babel/plugin-proposal-class-properties',
     ['@babel/plugin-transform-runtime', {
       'corejs': 2
-    }],
-    ['babel-plugin-transform-define', {
-      'process.env.NEXT_VERSION': require('./package.json').version
     }]
   ]
 }
