+++ lerna.json
@@ -1,4 +1,6 @@
 {
+  "npmClient": "yarn",
+  "useWorkspaces": true,
   "packages": [
     "packages/*"
   ],
