+++ packages/next/babel.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  'presets': [require('../../babel.config.js')],
+  'plugins': [
+    ['babel-plugin-transform-define', {
+      'process.env.NEXT_VERSION': require('./package.json').version
+    }]
+  ]
+}
