+++ packages/next/package.json
@@ -29,17 +29,9 @@
   "scripts": {
     "build": "taskr",
     "release": "taskr release",
-    "pretestonly": "taskr pretest",
-    "testonly": "cross-env NODE_PATH=test/lib jest \\.test.js",
-    "posttestonly": "taskr posttest",
-    "testall": "npm run testonly -- --coverage --forceExit --runInBand --verbose --bail",
-    "pretest": "npm run lint && npm run flow",
-    "test": "cross-env npm run testall || npm run testall",
-    "coveralls": "nyc --instrument=false --source-map=false report --temp-directory=./coverage --reporter=text-lcov | coveralls",
     "flow": "flow check",
     "lint": "standard 'bin/*' 'client/**/*.js' 'examples/**/*.js' 'lib/**/*.js' 'pages/**/*.js' 'server/**/*.js' 'build/**/*.js' 'test/**/*.js'",
-    "prepublish": "npm run release",
-    "precommit": "lint-staged"
+    "prepublish": "npm run release"
   },
   "taskr": {
     "requires": [
@@ -55,21 +47,17 @@
       "**/examples/with-mobx/**"
     ]
   },
-  "lint-staged": {
-    "*.js": "standard",
-    "bin/*": "standard"
-  },
   "dependencies": {
-    "@babel/core": "7.0.0",
-    "@babel/plugin-proposal-class-properties": "7.0.0",
+    "@babel/core": "7.1.2",
+    "@babel/plugin-proposal-class-properties": "7.1.0",
     "@babel/plugin-proposal-object-rest-spread": "7.0.0",
     "@babel/plugin-syntax-dynamic-import": "7.0.0",
-    "@babel/plugin-transform-runtime": "7.0.0",
-    "@babel/preset-env": "7.0.0",
+    "@babel/plugin-transform-runtime": "7.1.0",
+    "@babel/preset-env": "7.1.0",
     "@babel/preset-react": "7.0.0",
-    "@babel/runtime": "7.0.0",
-    "@babel/runtime-corejs2": "7.0.0",
-    "@babel/template": "7.0.0",
+    "@babel/runtime": "7.1.2",
+    "@babel/runtime-corejs2": "7.1.2",
+    "@babel/template": "7.1.2",
     "ansi-html": "0.0.7",
     "autodll-webpack-plugin": "0.4.2",
     "babel-core": "7.0.0-bridge.0",
@@ -114,52 +102,13 @@
     "webpackbar": "2.6.3",
     "write-file-webpack-plugin": "4.3.2"
   },
-  "devDependencies": {
-    "@babel/preset-flow": "7.0.0",
-    "@taskr/clear": "1.1.0",
-    "@taskr/esnext": "1.1.0",
-    "@taskr/watch": "1.1.0",
-    "@zeit/next-css": "0.2.1-canary.1",
-    "@zeit/next-sass": "0.2.1-canary.1",
-    "@zeit/next-typescript": "1.1.0",
-    "babel-eslint": "9.0.0",
-    "babel-jest": "23.4.2",
-    "babel-plugin-transform-define": "1.3.0",
-    "benchmark": "2.1.4",
-    "cheerio": "0.22.0",
-    "chromedriver": "2.42.0",
-    "clone": "2.1.1",
-    "coveralls": "2.13.1",
-    "cross-env": "5.2.0",
-    "express": "4.16.3",
-    "fkill": "5.1.0",
-    "flatten": "1.0.2",
-    "flow-bin": "0.73.0",
-    "get-port": "3.2.0",
-    "husky": "0.14.3",
-    "jest-cli": "21.2.0",
-    "lint-staged": "4.2.3",
-    "micro": "9.1.0",
-    "mkdirp": "0.5.1",
-    "node-fetch": "1.7.3",
-    "node-notifier": "5.1.2",
-    "node-sass": "4.9.2",
-    "nyc": "11.2.1",
-    "react": "16.4.2",
-    "react-dom": "16.4.2",
-    "rimraf": "2.6.2",
-    "standard": "11.0.1",
-    "taskr": "1.1.0",
-    "wd": "1.10.3"
-  },
   "peerDependencies": {
     "react": "^16.0.0",
     "react-dom": "^16.0.0"
   },
-  "jest": {
-    "testEnvironment": "node",
-    "roots": [
-      "test/"
-    ]
+  "devDependencies": {
+    "@babel/preset-flow": "7.0.0",
+    "@taskr/clear": "1.1.0",
+    "@taskr/watch": "1.1.0"
   }
 }
